requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
build==0.10.0
    # via pip-tools
click==8.1.3
    # via pip-tools
django==4.2.2
    # via
    #   -r requirements.in
    #   djangorestframework
django-rest-framework==0.1.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via django-rest-framework
packaging==23.1
    # via build
pip-tools==6.13.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
pytz==2023.3
    # via djangorestframework
sqlparse==0.4.4
    # via django
tomli==2.0.1
    # via build
typing-extensions==4.7.0rc1
    # via asgiref
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
